backend/requirements.txt
fastapi
google-generativeai
pydantic
uvicorn
pillow